Core blocks: Extract edit to their own file - part 1
