+++ core-blocks/button/index.js
@@ -8,157 +8,16 @@ import { omit, pick } from 'lodash';
  * WordPress dependencies
  */
 import { __ } from '@wordpress/i18n';
-import { Component, Fragment } from '@wordpress/element';
 import {
-	Dashicon,
-	IconButton,
-	PanelBody,
-	ToggleControl,
-	withFallbackStyles,
-} from '@wordpress/components';
-import {
-	UrlInput,
 	RichText,
-	BlockControls,
-	BlockAlignmentToolbar,
-	ContrastChecker,
-	InspectorControls,
 	getColorClass,
-	withColors,
-	PanelColor,
 } from '@wordpress/editor';
 
 /**
  * Internal dependencies
  */
-import './editor.scss';
 import './style.scss';
-
-const { getComputedStyle } = window;
-
-const ContrastCheckerWithFallbackStyles = withFallbackStyles( ( node, ownProps ) => {
-	const { textColor, backgroundColor } = ownProps;
-	//avoid the use of querySelector if textColor color is known and verify if node is available.
-	const textNode = ! textColor && node ? node.querySelector( '[contenteditable="true"]' ) : null;
-	return {
-		fallbackBackgroundColor: backgroundColor || ! node ? undefined : getComputedStyle( node ).backgroundColor,
-		fallbackTextColor: textColor || ! textNode ? undefined : getComputedStyle( textNode ).color,
-	};
-} )( ContrastChecker );
-
-class ButtonBlock extends Component {
-	constructor() {
-		super( ...arguments );
-		this.nodeRef = null;
-		this.bindRef = this.bindRef.bind( this );
-		this.updateAlignment = this.updateAlignment.bind( this );
-		this.toggleClear = this.toggleClear.bind( this );
-	}
-
-	updateAlignment( nextAlign ) {
-		this.props.setAttributes( { align: nextAlign } );
-	}
-
-	toggleClear() {
-		const { attributes, setAttributes } = this.props;
-		setAttributes( { clear: ! attributes.clear } );
-	}
-
-	bindRef( node ) {
-		if ( ! node ) {
-			return;
-		}
-		this.nodeRef = node;
-	}
-
-	render() {
-		const {
-			attributes,
-			backgroundColor,
-			textColor,
-			setBackgroundColor,
-			setTextColor,
-			setAttributes,
-			isSelected,
-			className,
-		} = this.props;
-
-		const {
-			text,
-			url,
-			title,
-			align,
-			clear,
-		} = attributes;
-
-		return (
-			<Fragment>
-				<BlockControls>
-					<BlockAlignmentToolbar value={ align } onChange={ this.updateAlignment } />
-				</BlockControls>
-				<span className={ className } title={ title } ref={ this.bindRef }>
-					<RichText
-						tagName="span"
-						placeholder={ __( 'Add text…' ) }
-						value={ text }
-						onChange={ ( value ) => setAttributes( { text: value } ) }
-						formattingControls={ [ 'bold', 'italic', 'strikethrough' ] }
-						className={ classnames(
-							'wp-block-button__link', {
-								'has-background': backgroundColor.value,
-								[ backgroundColor.class ]: backgroundColor.class,
-								'has-text-color': textColor.value,
-								[ textColor.class ]: textColor.class,
-							}
-						) }
-						style={ {
-							backgroundColor: backgroundColor.class ? undefined : backgroundColor.value,
-							color: textColor.class ? undefined : textColor.value,
-						} }
-						keepPlaceholderOnFocus
-					/>
-					<InspectorControls>
-						<PanelBody>
-							<ToggleControl
-								label={ __( 'Wrap text' ) }
-								checked={ !! clear }
-								onChange={ this.toggleClear }
-							/>
-							<PanelColor
-								colorValue={ backgroundColor.value }
-								title={ __( 'Background Color' ) }
-								onChange={ setBackgroundColor }
-							/>
-							<PanelColor
-								colorValue={ textColor.value }
-								title={ __( 'Text Color' ) }
-								onChange={ setTextColor }
-							/>
-							{ this.nodeRef && <ContrastCheckerWithFallbackStyles
-								node={ this.nodeRef }
-								textColor={ textColor.value }
-								backgroundColor={ backgroundColor.value }
-								isLargeText={ true }
-							/> }
-						</PanelBody>
-					</InspectorControls>
-				</span>
-				{ isSelected && (
-					<form
-						className="blocks-button__inline-link"
-						onSubmit={ ( event ) => event.preventDefault() }>
-						<Dashicon icon="admin-links" />
-						<UrlInput
-							value={ url }
-							onChange={ ( value ) => setAttributes( { url: value } ) }
-						/>
-						<IconButton icon="editor-break" label={ __( 'Apply' ) } type="submit" />
-					</form>
-				) }
-			</Fragment>
-		);
-	}
-}
+import edit from './edit';
 
 const blockAttributes = {
 	url: {
@@ -232,14 +91,7 @@ export const settings = {
 		return props;
 	},
 
-	edit: withColors( ( getColor, setColor, { attributes, setAttributes } ) => {
-		return {
-			backgroundColor: getColor( attributes.backgroundColor, attributes.customBackgroundColor, 'background-color' ),
-			setBackgroundColor: setColor( 'backgroundColor', 'customBackgroundColor', setAttributes ),
-			textColor: getColor( attributes.textColor, attributes.customTextColor, 'color' ),
-			setTextColor: setColor( 'textColor', 'customTextColor', setAttributes ),
-		};
-	} )( ButtonBlock ),
+	edit,
 
 	save( { attributes } ) {
 		const {
